Reorganize Control.Effect.State hierarchy.
@robrix pointed out that it's good practice to avoid prefixing data
types with their name, instead using modules and module-specific
imports to say what you mean. I had chosen the opposite in order to
have people opt-in twice to use lazy state (since lazy state is rarely
appropriate for most programs), but I think internal consistency
trumps a second opt-in.

The lazy and strict state carriers now live in distinct modules. Both
modules reexport the core `State` effect defined in an internal
module. This maintains backwards compatibility by making
`Control.Effect.State` an alias for `Control.Effect.State.Strict`. (I
had also hoped to avoid an internal module, but at this point it's by
far the simplest solution.)
