+++ src/Control/Effect/State/Lazy.hs
@@ -6,84 +6,79 @@ module Control.Effect.State.Lazy
 , put
 , modify
 , modifyLazy
-, runLazyState
-, evalLazyState
-, execLazyState
-, LazyStateC(..)
+, StateC(..)
+, runState
+, evalState
+, execState
 ) where
 
 import Control.Applicative (Alternative(..))
 import Control.Effect.Carrier
 import Control.Effect.Sum
-import Control.Effect.State
+import Control.Effect.State.Internal
 import Control.Monad (MonadPlus(..))
 import Control.Monad.Fail
 import Control.Monad.IO.Class
 import Control.Monad.Trans.Class
 import Prelude hiding (fail)
 
--- | Replace the state value with the result of applying a function to the current state value.
---   This is lazy in the new state; injudicious use of this function may lead to space leaks.
-modifyLazy :: (Member (State s) sig, Carrier sig m) => (s -> s) -> m ()
-modifyLazy f = get >>= put . f
+newtype StateC s m a = StateC { runStateC :: s -> m (s, a) }
 
-newtype LazyStateC s m a = LazyStateC { runLazyStateC :: s -> m (s, a) }
+instance Functor m => Functor (StateC s m) where
+  fmap f m = StateC $ \ s -> fmap (\ ~(s', a) -> (s', f a)) $ runStateC m s
 
-instance Functor m => Functor (LazyStateC s m) where
-  fmap f m = LazyStateC $ \ s -> fmap (\ ~(s', a) -> (s', f a)) $ runLazyStateC m s
-
-instance (Functor m, Monad m) => Applicative (LazyStateC s m) where
-  pure a = LazyStateC $ \ s -> pure (s, a)
-  LazyStateC mf <*> LazyStateC mx = LazyStateC $ \ s -> do
+instance (Functor m, Monad m) => Applicative (StateC s m) where
+  pure a = StateC $ \ s -> pure (s, a)
+  StateC mf <*> StateC mx = StateC $ \ s -> do
     ~(s', f) <- mf s
     ~(s'', x) <- mx s'
     return (s'', f x)
 
-instance Monad m => Monad (LazyStateC s m) where
-  m >>= k  = LazyStateC $ \ s -> do
-    ~(s', a) <- runLazyStateC m s
-    runLazyStateC (k a) s'
+instance Monad m => Monad (StateC s m) where
+  m >>= k  = StateC $ \ s -> do
+    ~(s', a) <- runStateC m s
+    runStateC (k a) s'
 
-instance (Alternative m, Monad m) => Alternative (LazyStateC s m) where
-  empty = LazyStateC (const empty)
-  LazyStateC l <|> LazyStateC r = LazyStateC (\ s -> l s <|> r s)
+instance (Alternative m, Monad m) => Alternative (StateC s m) where
+  empty = StateC (const empty)
+  StateC l <|> StateC r = StateC (\ s -> l s <|> r s)
 
-instance MonadFail m => MonadFail (LazyStateC s m) where
-  fail s = LazyStateC (const (fail s))
+instance MonadFail m => MonadFail (StateC s m) where
+  fail s = StateC (const (fail s))
 
-instance MonadIO m => MonadIO (LazyStateC s m) where
-  liftIO io = LazyStateC (\ s -> (,) s <$> liftIO io)
+instance MonadIO m => MonadIO (StateC s m) where
+  liftIO io = StateC (\ s -> (,) s <$> liftIO io)
 
-instance (Alternative m, Monad m) => MonadPlus (LazyStateC s m)
+instance (Alternative m, Monad m) => MonadPlus (StateC s m)
 
-instance MonadTrans (LazyStateC s) where
-  lift m = LazyStateC (\ s -> (,) s <$> m)
+instance MonadTrans (StateC s) where
+  lift m = StateC (\ s -> (,) s <$> m)
 
-instance (Carrier sig m, Effect sig) => Carrier (State s :+: sig) (LazyStateC s m) where
-  eff (L (Get   k)) = LazyStateC (\ s -> runLazyState s (k s))
-  eff (L (Put s k)) = LazyStateC (\ _ -> runLazyState s k)
-  eff (R other)     = LazyStateC (\ s -> eff (handle (s, ()) (uncurry runLazyState) other))
+instance (Carrier sig m, Effect sig) => Carrier (State s :+: sig) (StateC s m) where
+  eff (L (Get   k)) = StateC (\ s -> runState s (k s))
+  eff (L (Put s k)) = StateC (\ _ -> runState s k)
+  eff (R other)     = StateC (\ s -> eff (handle (s, ()) (uncurry runState) other))
 
 -- | Run a lazy 'State' effect, yielding the result value and the final state.
 --   More programs terminate with lazy state than strict state, but injudicious
 --   use of lazy state may lead to thunk buildup.
 --
---   prop> run (runLazyState a (pure b)) == (a, b)
---   prop> take 5 . snd . run $ runLazyState () (traverse pure [1..]) == [1,2,3,4,5]
-runLazyState :: s -> LazyStateC s m a -> m (s, a)
-runLazyState = flip runLazyStateC
+--   prop> run (runState a (pure b)) == (a, b)
+--   prop> take 5 . snd . run $ runState () (traverse pure [1..]) == [1,2,3,4,5]
+runState :: s -> StateC s m a -> m (s, a)
+runState = flip runStateC
 
 -- | Run a lazy 'State' effect, yielding the result value and discarding the final state.
 --
---   prop> run (evalLazyState a (pure b)) == b
-evalLazyState :: forall s m a . Functor m => s -> LazyStateC s m a -> m a
-evalLazyState s = fmap snd . runLazyState s
+--   prop> run (evalState a (pure b)) == b
+evalState :: forall s m a . Functor m => s -> StateC s m a -> m a
+evalState s = fmap snd . runState s
 
 -- | Run a lazy 'State' effect, yielding the final state and discarding the return value.
 --
---   prop> run (execLazyState a (pure b)) == a
-execLazyState :: forall s m a . Functor m => s -> LazyStateC s m a -> m s
-execLazyState s = fmap fst . runLazyState s
+--   prop> run (execState a (pure b)) == a
+execState :: forall s m a . Functor m => s -> StateC s m a -> m s
+execState s = fmap fst . runState s
 
 -- $setup
 -- >>> :seti -XFlexibleContexts
